chore: address RBAC review comments
