+++ server/internal/access/queries.sql
@@ -61,19 +61,6 @@ DELETE FROM principal_grants
 WHERE organization_id = @organization_id
   AND principal_urn = @principal_urn;
 
--- name: HasActiveOrganizationUser :one
--- Returns whether a Gram user is an active member of the organization.
-SELECT EXISTS(
-  SELECT 1
-  FROM users
-  JOIN organization_user_relationships
-    ON organization_user_relationships.user_id = users.id
-  WHERE users.id = @user_id
-    AND users.deleted_at IS NULL
-    AND organization_user_relationships.organization_id = @organization_id
-    AND organization_user_relationships.deleted_at IS NULL
-) AS exists;
-
 -- Queries for authz challenge resolutions.
 -- authz_challenge_resolutions is org-scoped (no project_id).
