+++ server/internal/authz/resource_kinds.go
@@ -1,0 +1,10 @@
+package authz
+
+const (
+	ResourceKindWildcard    = WildcardResource
+	ResourceKindProject     = "project"
+	ResourceKindMCP         = "mcp"
+	ResourceKindOrg         = "org"
+	ResourceKindEnvironment = "environment"
+	ResourceKindRiskPolicy  = "risk_policy"
+)
